Create POSCart to contain POSCartItem and POSCartCouponItem
